ci: :zap: Adds GitHub Actions pipeline for Django with linters, documentation generation, security analysis, and code quality checks
Add GitHub Actions pipeline for Django  This commit introduces a GitHub Actions workflow to automate various tasks related to the development of our Django application. The CI/CD pipeline includes the following actions:  - Initial configuration of the workflow to run on every push and pull request to the main branch. - Installation of dependencies and configuration of the Python environment. - Running linters (flake8, pylint) to verify compliance with style rules and code conventions. - Automatic generation of documentation using Sphinx. - Scanning dependencies for known vulnerabilities using Safety. - Verification of code coverage using Coverage and sending a report to Codecov. - Static analysis of the code in search of security and quality problems using Bandit. - Scanning code for possible secrets or sensitive information using detect-secrets and git-secrets.  This workflow improves code quality and security, ensures documentation is always up-to-date, and provides automatic code review for every change.